mogadm: add "device next" sub-command
This is a convenience command to ease the process of adding devices
to a cluster.  This may be useful when there are many devices to be
added.  This provides no synchronization and multiple administrative
users will still have to coordinate device additions to avoid errors
on duplicates.

Adding synchronization would require extending the tracker
protocol, and is probably not worth the complexity.